refactor: revert HotContainer for now
