+++ src/index.prod.js
@@ -1,2 +1,1 @@
-module.exports.HotContainer = require('./HotContainer')
-module.exports.AppContainer = require('./HotContainer')
+module.exports.AppContainer = require('./AppContainer')
